optimising code
