+++ src/components/Domain/DomainType.vue
@@ -21,12 +21,12 @@ import {mapActions, mapGetters, mapState} from "vuex"
 import * as am5 from '@amcharts/amcharts5';
 import * as am5hierarchy from "@amcharts/amcharts5/hierarchy";
 import am5themes_Animated from '@amcharts/amcharts5/themes/Animated';
-import { canvasGetImageData } from "@/utils/canvasRenderingContext"
-import { breadCrumbBar } from "@/utils/breadCrumbBar"
-import StringMixin from "@/utils/stringMixin.js"
-import RecordTypes from "@/utils/recordTypes.js";
+import { canvasGetImageData } from "@/utils/Others/canvasRenderingContext"
+import { breadCrumbBar } from "@/utils/Others/breadCrumbBar"
+import StringMixin from "@/utils/Others/stringMixin.js"
+import RecordTypes from "@/utils/Others/recordTypes.js";
 import Loaders from "@/components/Loaders/Loaders"
-import calculateRecords from "@/utils/calculateRecords";
+import calculateRecords from "@/utils/Others/calculateRecords";
 
 export default {
   name: 'DomainType',
