+++ src/components/Domain/DomainType_D3JS_1.vue
@@ -105,7 +105,7 @@ export default {
         update(d);
       }
 
-      function update() {
+      function updateTick() {
         var nodes = flatten(root),
           links = d3.layout.tree().links(nodes);
