--- src/data/index.js
@@ -1,9 +1,0 @@
-import resourcetype from './resourcetype.json'
-import sunburstdata from './sunburst.json'
-import homePageData from "./homePageData.json"
-
-export {
-  resourcetype,
-  sunburstdata,
-  homePageData,
-}
