--- src/data/sunburst.json
@@ -1,57 +1,0 @@
-{
-  "name": "Root",
-  "records_count": 30,
-  "children": [{
-    "name": "Natural Science",
-    "records_count": 30,
-    "children": [{
-      "name": "A00",
-      "records_count": 71
-    }, {
-      "name": "A01",
-      "records_count": 33
-    }, {
-      "name": "A02",
-      "records_count": 2
-    }]
-  }, {
-    "name": "Humanities and Social Science",
-    "records_count": 50,
-    "children": [{
-      "name": "B10",
-      "records_count": 78
-    }, {
-      "name": "B11",
-      "records_count": 36
-    }]
-  },
-  {
-  "name": "Subject Agnostic",
-    "records_count": 50,
-  "children": [{
-    "name": "C20",
-    "records_count": 46
-  }, {
-    "name": "C21",
-    "records_count": 28
-  }, {
-    "name": "C22",
-    "records_count": 18
-  }]
-  },
-    {
-      "name": "Engineering Science",
-      "records_count": 60,
-      "children": [{
-        "name": "C20",
-        "records_count": 46
-      }, {
-        "name": "C21",
-        "records_count": 28
-      }, {
-        "name": "C22",
-        "records_count": 18
-      }]
-    }
-  ]
-}
