+++ src/lib/D3GraphClient/SubjectGraphUtils/index.js
@@ -1,0 +1,9 @@
+import {fillColor, linkColor, strokeColor} from "./subjectColors"
+import nodeSize from "./subjectNodeSize"
+
+export {
+  fillColor,
+  linkColor,
+  strokeColor,
+  nodeSize
+}
