+++ src/lib/D3GraphClient/addClass.js
@@ -1,0 +1,7 @@
+// Add class to node
+const addClass = (d)=> {
+  if (d._children?.length || d.children?.length) return "node"
+  return "node noChlid"
+}
+
+export default addClass
