+++ src/lib/D3GraphClient/parseLevel.js
@@ -1,0 +1,10 @@
+const parseLevel = (node, level) => {
+  node.level = level;
+  if (node.children && node.children.length) {
+    node.children.forEach((children) => {
+      parseLevel(children, level + 1);
+    });
+  }
+}
+
+export default parseLevel
