+++ src/utils/Others/canvasRenderingContext.js
