+++ src/utils/Others/stringMixin.js
